fix(tests): Move linux tests to dedicated subfolder
