+++ test/new-e2e/tests/installer/unix/all_packages.go
