+++ test/new-e2e/tests/installer/unix/all_packages_test.go
