+++ test/new-e2e/tests/installer/unix/package_agent_test.go
