+++ test/new-e2e/tests/installer/unix/package_apm_inject_test.go
