+++ test/new-e2e/tests/installer/unix/package_definitions.go
