+++ test/new-e2e/tests/installer/unix/package_installer_test.go
