+++ test/new-e2e/tests/installer/unix/upgrade_scenario_test.go
